# feat: logger inheritance