--- a/tests/functional/test_aws_lambda_logging.py
+++ b/tests/functional/test_aws_lambda_logging.py
@@ -1,11 +1,10 @@
 """aws_lambda_logging tests."""
 import io
 import json
-import logging
 
 import pytest
 
-from aws_lambda_powertools import Logger
+from .utility_functions import get_random_logger
 
 
 @pytest.fixture
@@ -13,26 +12,9 @@ def stdout():
     return io.StringIO()
 
 
-@pytest.fixture
-def handler(stdout):
-    return logging.StreamHandler(stdout)
-
-
-@pytest.fixture
-def logger():
-    return logging.getLogger(__name__)
-
-
-@pytest.fixture
-def root_logger(handler):
-    logging.root.addHandler(handler)
-    yield logging.root
-    logging.root.removeHandler(handler)
-
-
 @pytest.mark.parametrize("level", ["DEBUG", "WARNING", "ERROR", "INFO", "CRITICAL"])
-def test_setup_with_valid_log_levels(root_logger, stdout, level):
-    logger = Logger(level=level, stream=stdout, request_id="request id!", another="value")
+def test_setup_with_valid_log_levels(stdout, level):
+    logger = get_random_logger(level=level, stream=stdout, request_id="request id!", another="value")
     msg = "This is a test"
     log_command = {
         "INFO": logger.info,
@@ -55,8 +37,8 @@ def test_setup_with_valid_log_levels(root_logger, stdout, level):
     assert "exception" not in log_dict
 
 
-def test_logging_exception_traceback(root_logger, stdout):
-    logger = Logger(level="DEBUG", stream=stdout, request_id="request id!", another="value")
+def test_logging_exception_traceback(stdout):
+    logger = get_random_logger(level="DEBUG", stream=stdout, request_id="request id!", another="value")
 
     try:
         raise Exception("Boom")
@@ -69,9 +51,9 @@ def test_logging_exception_traceback(root_logger, stdout):
     assert "exception" in log_dict
 
 
-def test_setup_with_invalid_log_level(root_logger, logger, stdout):
+def test_setup_with_invalid_log_level(stdout):
     with pytest.raises(ValueError) as e:
-        Logger(level="not a valid log level")
+        get_random_logger(level="not a valid log level")
         assert "Unknown level" in e.value.args[0]
 
 
@@ -82,12 +64,8 @@ def check_log_dict(log_dict):
     assert "message" in log_dict
 
 
-def test_setup_with_bad_level_does_not_fail():
-    Logger("DBGG", request_id="request id!", another="value")
-
-
-def test_with_dict_message(root_logger, stdout):
-    logger = Logger(level="DEBUG", stream=stdout)
+def test_with_dict_message(stdout):
+    logger = get_random_logger(level="DEBUG", stream=stdout)
 
     msg = {"x": "isx"}
     logger.critical(msg)
@@ -97,8 +75,8 @@ def test_with_dict_message(root_logger, stdout):
     assert msg == log_dict["message"]
 
 
-def test_with_json_message(root_logger, stdout):
-    logger = Logger(stream=stdout)
+def test_with_json_message(stdout):
+    logger = get_random_logger(stream=stdout)
 
     msg = {"x": "isx"}
     logger.info(json.dumps(msg))
@@ -108,8 +86,8 @@ def test_with_json_message(root_logger, stdout):
     assert msg == log_dict["message"]
 
 
-def test_with_unserialisable_value_in_message(root_logger, stdout):
-    logger = Logger(level="DEBUG", stream=stdout)
+def test_with_unserialisable_value_in_message(stdout):
+    logger = get_random_logger(level="DEBUG", stream=stdout)
 
     class X:
         pass