**Issue #, if available:** #97 

## Description of changes:

<!--- One or two sentences as a summary of what's being changed -->

This addresses a common issue when trying to use `Logger` across multiple modules while wanting to:

* **Create a new child Logger** to inherit parent Logger's configuration (handler, formatter, custom keys, sampling)
* **Reuse the same Logger**, where Logger is enriched iteratively across the code base

For this, we've introduced a new parameter, `child`, when initializing a new Logger, and some internal changes while keeping the same UX and backwards compatibility.

* **[Internal]** Direct access to methods, handlers, etc work as before, except that we proxy properties access to `._logger`

## UX

**Key aspects**:

* We support logging inheritance via the `child` parameter, and the `Logger` will be named after `<service>.<caller_file_name>`
* Calling `Logger` with the same `service` name (explicit or via environment variable) returns the same `Logger`
* Changes to either Child `Logger` or Parent `Logger` will be propagated to all instances

### Creating a child Logger

Child logger can be created in any other as long as a parent is created too - This supports customers using Layers or different file names to create child loggers.

```python
# lambda_handler.py
from aws_lambda_powertools import Logger
import collect

logger = Logger(service="payment") # parent Logger just like before, named 'payment'
...

# collect.py
from aws_lambda_powertools import Logger

logger = Logger(service="payment", child=True) # child Logger named 'payment.collect'

# a_new_key will now be available both in this child logger as well as the parent
logger.structure_logs(append=True, a_new_key="value") 
...
```

### Reusing the same Logger

```python
# lambda_handler.py
from aws_lambda_powertools import Logger
import collect

logger = Logger(service="payment") # parent Logger just like before, named 'payment'

# collect.py
from aws_lambda_powertools import Logger

logger = Logger(service="payment") # returns the existing 'payment' Logger
...
```

**Checklist**

<!--- Leave unchecked if your change doesn't seem to apply --> 

- [x] [Meet tenets criteria](https://awslabs.github.io/aws-lambda-powertools-python/#tenets)
- [x] Update tests
- [ ] Update docs
- [x] PR title follows [conventional commit semantics](https://github.com/awslabs/aws-lambda-powertools-python/blob/376ec0a2ac0d2a40e0af5717bef42ff84ca0d1b9/.github/semantic.yml#L2)

**Changes**

- [x] Register logger via `getLogger` to allow inheritance
- [x] Encapsulate logger initialization/config in a new method
- [x] Remove Logger sub-classing
  - [x] Proxy calls to underneath Logger instead
- [x] ~~Create `name` parameter for hierarchy~~
  - [x] ~~Use `service` if `name` is empty~~
- [x] Test on Lambda locally
  - [x] Single Logger, no code changes
  - [x] Two Loggers in different files
  - [x] Additional Logger in a package that is imported on main entry-point
  - [x] Lambda Layer
- [x] Only add a handler if there isn't one
- [x] Only add a handler if parent is root
- [x] Update Formatter to append new keys
- [x] Make `handler` private
- [x] Remove `log_keys`
- [x] Remove duplicate Parameters in docstring
- [x] Double check Debug w/ getattr perf impact
- [x] Docs: Document inheriting Loggers
- [x] Docs: Changelog (doc random logger for tests)
- [x] Test: Logger hierarchy via logger.parent
- [x] Remove `name` parameter
- [x] Allow logger child propagation
- [x] Introduce `child` boolean param
- [x] Don't add a handler nor evaluate sampling if it's a `child` Logger
- [x] Name `child` Loggers based on `<service>.<caller_file_name>` to mimick `getLogger(__name__)`
- [x] Update tests to remove Handler on every test
- [x] Update tests to remove `getRandomLogger` utility function

**Optional**

- [x] Ensure log level input is case insensitive

## Benchmark

> **TL;DR**: For the average use case we have a slight performance decrease, but all negligible in all cases (0.10 nano seconds). 

Quick checks as we've removed `logging.Logger` sub-classing, respect Logging inheritance, and are proxying calls to inner Logger.

### Basic logger

* **Script**: `python -m timeit -s 'from aws_lambda_powertools import Logger; logger = Logger(); logger.info("hello world")'`
* **Before**: 50000000 loops, best of 5: **6.12 nsec** per loop
* **After with Logger proxy object, no sub-classing**: 50000000 loops, best of 5: **6.22 nsec** per loop

### Inheritance

* **Script**: `python -m timeit -s 'from aws_lambda_powertools import Logger; logger = Logger(); l = Logger(); l.info("hello world")'`
* **Same logger**: 50000000 loops, best of 5: **6.28 nsec** per loop
* **Random child logger name + fixup**: 50000000 loops, best of 5: **6.14 nsec** per loop
    - **Script**: `python -m timeit -s 'from aws_lambda_powertools import Logger; logger = Logger(child=True); l = Logger(); l.info("hello world")'`


## Breaking change checklist

<!--- Ignore if it's not a breaking change -->

**RFC issue #**:

* [ ] Migration process documented
* [ ] Implement warnings (if it can live side by side)

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.