@heitorlessa 2020-08-12
Just addressed @alexanderluiscampino commentson consistency, and took the opportunity to turn`log level` input to case insensitive to prevent common mistakes for newcomers.

I'll start working on the docs this afternoon or tomorrow, and should release a minor version by EOW