@heitorlessa 2020-08-12
Yep it’s fixed in the latest commit ;)

On Wed, 12 Aug 2020 at 09:32, Alexandre Campino <notifications@github.com>
wrote:

> *@alexanderluiscampino* commented on this pull request.
> ------------------------------
>
> In aws_lambda_powertools/logging/logger.py
> <https://github.com/awslabs/aws-lambda-powertools-python/pull/99#discussion_r469061718>
> :
>
> > @@ -118,7 +118,7 @@ def __init__(
>      ):
>          self.service = service or os.getenv("POWERTOOLS_SERVICE_NAME") or "service_undefined"
>          self.sampling_rate = sampling_rate or os.getenv("POWERTOOLS_LOGGER_SAMPLE_RATE") or 0.0
> -        self.log_level = level or os.getenv("LOG_LEVEL") or logging.INFO
> +        self.log_level = level or os.getenv("LOG_LEVEL".upper()) or logging.INFO
>
> Gotcha, your explanation makes sense, then you need the .upper outside of
> the getenv(LOG_LEVEL).UPPER()
>
> —
> You are receiving this because you were assigned.
> Reply to this email directly, view it on GitHub
> <https://github.com/awslabs/aws-lambda-powertools-python/pull/99#discussion_r469061718>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAZPQBH7FMGNN2SRN76WK2TSAJAQDANCNFSM4PZATT2Q>
> .
>